# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Biu IM API",
    "description": "Biu 即时通讯应用 REST API 文档",
    "version": "1.0.0",
    "contact": {
      "name": "Biu Dev Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发服务器"
    }
  ],
...
  "paths": {
...
    "/api/ai-roles/config/test": {
      "post": {
        "tags": [
          "ai-role"
        ],
        "summary": "[AI 角色与模型] 测试 AI 连接",
        "description": "需要认证 (Bearer Token)",
        "operationId": "post_modelConfigController_testConfigConnection",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/ai-roles/models/{id}/test": {
      "post": {
        "tags": [
          "ai-role"
        ],
        "summary": "[AI 模型管理] 测试模型连接",
        "description": "需要认证 (Bearer Token)",
        "operationId": "post_controller_testModel",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "使用登录接口获取的 JWT Token"
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "HTTP 状态码"
          },
          "message": {
            "type": "string",
            "description": "错误描述"
          },
          "details": {
            "type": "string",
            "description": "详细错误信息"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "HTTP 状态码"
          },
          "message": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "description": "业务数据"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}